requirements.txt
# Azure Fabric Management
azure-mgmt-fabric>=1.0.0
azure-identity>=1.15.0

# Azure Management SDKs (used by post-provision scripts)
azure-mgmt-search>=9.1.0
azure-mgmt-authorization>=4.0.0

# Azure AI Foundry — Agent SDK + Agent Framework
azure-ai-projects>=1.0.0b7
azure-ai-agents>=1.1.0b3
agent-framework[azure] --pre

# OneLake / Azure Storage
azure-storage-file-datalake>=12.14.0

# Snowflake
snowflake-connector-python>=3.6.0

# Data processing
pandas>=2.1.0
pyarrow>=14.0.0
pydantic>=2.5.0

# Unstructured data extraction
azure-ai-documentintelligence>=1.0.0

# Confluence
atlassian-python-api>=3.41.0
beautifulsoup4>=4.12.0

# HTTP
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
click>=8.1.0
